// src/app/Interfaces/review_interface.ts
export interface IReview {

    ReviewAltText: string;
    OrganiztionId: string;
    WebSiteName: string;
    id: string;
    ReviewName: string;
    ReviewNewURL: string;
    ReviewOldURL: string;
    ReviewSchema: string;
    ReviewDescription: string;
    ReviewMetaDataTitle: string;
    ReviewMetaDataKeywords: string;
    ReviewMetaDataDescription: string;
    Featured_Review_Flag: string;
    ReviewImageURL: string;
    ReviewImageURL3: string;
    ReviewImageURL4: string;
    ReviewImageURL2: string;
    ReviewImageURL1: string;
    DelFlag: string;

}